@@ -1,27 +1,31 @@
 {
-  "_from": "raw-body@2.3.3",
-  "_id": "raw-body@2.3.3",
+  "_from": "raw-body@2.4.0",
+  "_id": "raw-body@2.4.0",
   "_inBundle": false,
-  "_integrity": "sha512-9esiElv1BrZoI3rCDuOuKCBRbuApGGaDPQfjSflGxdy4oyzqghxu6klEkkVIvBje+FF0BX9coEv8KqW6X/7njw==",
+  "_integrity": "sha512-4Oz8DUIwdvoa5qMJelxipzi/iJIi40O5cGV1wNYp5hvZP8ZN0T+jiNkL0QepXs+EsQ9XJ8ipEDoiH70ySUJP3Q==",
   "_location": "/raw-body",
-  "_phantomChildren": {},
+  "_phantomChildren": {
+    "depd": "1.1.2",
+    "inherits": "2.0.3",
+    "toidentifier": "1.0.0"
+  },
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "raw-body@2.3.3",
+    "raw": "raw-body@2.4.0",
     "name": "raw-body",
     "escapedName": "raw-body",
-    "rawSpec": "2.3.3",
+    "rawSpec": "2.4.0",
     "saveSpec": null,
-    "fetchSpec": "2.3.3"
+    "fetchSpec": "2.4.0"
   },
   "_requiredBy": [
     "/body-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.3.3.tgz",
-  "_shasum": "1b324ece6b5706e153855bc1148c65bb7f6ea0c3",
-  "_spec": "raw-body@2.3.3",
-  "_where": "C:\\repositorios\\dwavisit\\backend\\node_modules\\body-parser",
+  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.4.0.tgz",
+  "_shasum": "a1ce6fb9c9bc356ca52e89256ab59059e13d0332",
+  "_spec": "raw-body@2.4.0",
+  "_where": "C:\\GIT\\dwa\\backend\\node_modules\\body-parser",
   "author": {
     "name": "Jonathan Ong",
     "email": "me@jongleberry.com",
@@ -42,24 +46,24 @@
     }
   ],
   "dependencies": {
-    "bytes": "3.0.0",
-    "http-errors": "1.6.3",
-    "iconv-lite": "0.4.23",
+    "bytes": "3.1.0",
+    "http-errors": "1.7.2",
+    "iconv-lite": "0.4.24",
     "unpipe": "1.0.0"
   },
   "deprecated": false,
   "description": "Get and validate the raw body of a readable stream.",
   "devDependencies": {
-    "bluebird": "3.5.1",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-standard": "3.1.0",
+    "bluebird": "3.5.4",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3",
+    "mocha": "6.1.3",
     "readable-stream": "2.3.6",
     "safe-buffer": "5.1.2"
   },
@@ -86,5 +90,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --trace-deprecation --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --trace-deprecation --reporter spec --check-leaks test/"
   },
-  "version": "2.3.3"
+  "version": "2.4.0"
 }
