@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
+import { FormsModule } from '@angular/forms'
 import { MatCardModule, MatIconModule, MatToolbarModule , MatListModule, MatGridListModule, MatDividerModule } from '@angular/material';
 import { MatTableModule, MatButtonModule, MatInputModule, MatRadioModule } from '@angular/material';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
@@ -13,11 +14,12 @@ import { MoradoresComponent } from './panel/moradores/moradores.component';
 import { VisitantesComponent } from './panel/visitantes/visitantes.component';
 import { routing } from './app.routing';
 import { InfoComponent } from './panel/info/info.component';
-import { TableBasicExampleComponent } from './components/lista/table-basic-example';
+import { ListaMoradoresComponent } from './components/lista/lista-moradores/lista-moradores.component';
 import { CadastroMoradoresComponent } from './cadastro/cadastro-moradores/cadastro-moradores.component';
 import { ListaVisitantesComponent } from './components/lista/lista-visitantes/lista-visitantes.component';
 import { CadastroVisitantesComponent } from './cadastro/cadastro-visitantes/cadastro-visitantes.component';
-
+import { DialogEditMoradorComponent } from './components/dialog-edit-morador/dialog-edit-morador.component';
+import { MatDialogModule } from '@angular/material';
 
 @NgModule({
   declarations: [
@@ -29,10 +31,11 @@ import { CadastroVisitantesComponent } from './cadastro/cadastro-visitantes/cada
     MoradoresComponent,
     VisitantesComponent,
     InfoComponent,
-    TableBasicExampleComponent,
+    ListaMoradoresComponent,
     CadastroMoradoresComponent,
     ListaVisitantesComponent,
-    CadastroVisitantesComponent
+    CadastroVisitantesComponent,
+    DialogEditMoradorComponent
   ],
   imports: [
     BrowserModule,
@@ -48,7 +51,12 @@ import { CadastroVisitantesComponent } from './cadastro/cadastro-visitantes/cada
     MatButtonModule,
     MatInputModule,
     MatRadioModule,
-    HttpClientModule
+    HttpClientModule,
+    FormsModule,
+    MatDialogModule
+  ],
+  entryComponents: [
+    DialogEditMoradorComponent
   ],
   providers: [],
   bootstrap: [AppComponent]
