@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ParsedEvent } from '@angular/compiler';
 import { JsonpCallbackContext } from '@angular/common/http/src/jsonp';
+import { FormBuilder, FormControl, FormGroup, FormGroupDirective, NgForm } from '@angular/forms';
+import { MongoService } from 'src/app/services/mongo.service';
+import { Morador } from 'src/app/panel/moradores/morador.model';
 
 @Component({
   selector: 'dwa-cadastro-moradores',
@@ -10,17 +13,45 @@ import { JsonpCallbackContext } from '@angular/common/http/src/jsonp';
 })
 export class CadastroMoradoresComponent implements OnInit {
   http: HttpClient;
-  foto: any;
-  constructor() { }
+  FormMorador: FormGroup;
+
+  possuiVeiculo: string;
+  optionsVeiculo: string [] = ['Sim', 'Não'];
+  morador: Morador = {
+    _id: '',
+    apartamento: 0,
+    bloco: '',
+    cpf: 0,
+    email: '',
+    image: '',
+    nome: '',
+    possui_veiculo: false,
+    placa_veiculo: '',
+    modelo_veiculo: '',
+    cor_veiculo: '',
+    telefone: ''
+  };
+  // tslint:disable-next-line:no-shadowed-variable
+  constructor(private MongoService: MongoService) { }
 
   ngOnInit() {
   }
 
-  // TestaAPI() {
-  //   const data = this.http.get('http://countryapi.gear.host/v1/Country/getCountries').subscribe(dado => {
-  //   this.foto = dado;
-  //   console.log(this.foto);
-  // });
-
+  addMorador(morador: Morador) {
+    console.log(morador);
+    // const hab: Morador = {nome, cpf, bloco, apartamento, telefone, email,
+    //   possui_veiculo: (this.possuiVeiculo === 'Não' ? false : true), image};
+    // this.MongoService.postMorador(morador).subscribe(
+    //   // tslint:disable-next-line:no-shadowed-variable
+    //   (morador: Morador) => {
+    //     alert('Morador cadastrado com sucesso!');
+    //     console.log(morador);
+    //   },
+    //   (err) => {
+    //     alert('Ocorreu um erro ao salvar o Morador!');
+    //     console.log(err);
+    //   }
+    // );
+  }
 
 }
